Remove local absolute path from jflex_generate.sh
jflex_generate.sh was hardcoded to the location of the
IDEA jflex tools on my local machine.  Not so friendly for
contributors...

The script now takes the path to the root of the IDEA
source code as an argument.  Developers may also set
env var IDEA_SRC_ROOT to use the script with no
arguments.