Fix issue in EnterHandler and refactor ActionHandler tests
Fix an issue where if the caret is at the end of the file and the user presses enter after an open block tag,
we error trying to find a close block tag past the end of the file.  Also added testing around
this functionality.

While adding testing around this, refactor tests to make it easier to write action handler tests.