Fix issue in code to adjust formatting when a close tag is typed
Replaced the call to AutoIndentLinesHandler with a direct appeal to
codeStyleManager.adjustLineIndent adjust the formatting when an end
tag is typed.  This skips over issues of misplacing the caret
after the format (since AutoIndentLinesHandler insists on
kicking it to the next line).

Also got all editor.action and formatter tests passing