Highlight "else" expressions more intuitively
The previous behavior lexed "{{else" as an OPEN_INVERSE token.
While technically correct, this led to the undesirable behavior
of "else" being highlighted like "{{" rather than the same as,
say, "if".

Modify the lexer to produce more refined tokens in this
situation and assign the "identifier"-type highlight to the
resulting ELSE token