Make Commenter Language configurable
Add an option to select which language's commenter to use for the
"Comment with Line Comment" and "Comment with Block Comment" actions
in our templates.

This is needed since it doesn't look like there's a way to pass our
commenter context when it is invoked, so don't know what language we're
templating (and there's no proper comment construct in mustache).
This at least gives the user the option to have correct comments
for their most used template data language.