Implement structure view names and icons
Associates a logical name with each node in the structure view tree,
and adds an initial set of icons (these still need a bit of work).

Also builds out the Psi tree a bit to support this, cleaning up some of
the existing classes and adding elements for Paths ("foo/bar/baz") and
Data ("@data").